PaulsHappyPlace/requirements.txt
matplotlib==3.7.1
numpy==1.25.0
Pillow==10.1.0
scikit_learn==1.3.2
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
urllib3==2.0.7
pprint